Finishgit add .

+++ src/main/java/org/example/service/UserService.java
@@ -7,11 +7,11 @@ import org.example.model.User;
 public interface UserService {
     User addUser(User user);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 
     List<User> getAll();
 
-    User getById(Integer id);
+    User getById(Long id);
 
     void deleteAll();
